+++ testing/bulk_async/cuda/device_lambda/void_result.cu
@@ -22,7 +22,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents);
   }
@@ -44,7 +44,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents * 13);
   }
@@ -66,7 +66,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents * 13);
   }
